adding Front

+++ src/main/java/commu_bas/board/service/BoardDeleteReg.java
@@ -22,32 +22,26 @@ public class BoardDeleteReg implements BoardService {
 		}
 
 		BoardDTO dto = new BoardDTO();
+		
 		String post_id = request.getParameter("post_id");
+		String deleteCheck = request.getParameter("deleteCheck");
+		
 		dto.setPost_id(post_id);
-		dto.setPw(request.getParameter("pw"));
-
-		BoardDTO delDTO = new BoardDAO().detail(post_id);
-
-		int res = new BoardDAO().delete(dto);
-
-		String msg = "삭제 실패", goUrl = "DeleteForm?post_id=" + dto.getPost_id();
-
-		if (res > 0) {
-
-			msg = "삭제 성공";
-			goUrl = "List" + "?page=" + request.getParameter("nowPage");
-
-			if (delDTO.getUpfile() == null) {
-				String path = request.getRealPath("board");
-				path = "/Users/minsookim/Desktop/프로젝트/04_proj/proj_04_minsoo/proj_04_minsoo/src/main/webapp/uploadFile/commu_bas/board";
-				new File(path + "/" + delDTO.getUpfile()).delete();
-			}
+		int res = 0;
+		if(deleteCheck.equals("삭제")) {
+			res = new BoardDAO().delete(dto);
+			System.out.println(dto);
 		}
-
+		 
+		String msg = "삭제 실패", goUrl = "DeleteForm?post_id="+dto.getPost_id()+"&page="+request.getAttribute("nowPage");
+		
+		if(res>0) {
+			msg = "삭제성공";
+			goUrl = "List?page="+request.getAttribute("nowPage");
+		}
+		
 		request.setAttribute("msg", msg);
 		request.setAttribute("goUrl", goUrl);
 		request.setAttribute("mainUrl", "commu_bas/board/alert");
-		System.out.println("BoardModifyReg execute() 실행");
-
 	}
 }
